Clean up E2E caching
