+++ src/Backend/Program.cs
@@ -28,7 +28,7 @@ builder.Services.AddScoped<TicketSummarizer>();
 builder.Services.AddHttpClient<PythonInferenceClient>(c => c.BaseAddress = new Uri("http://python-inference"));
 builder.AddAzureBlobClient("eshopsupport-blobs");
 
-builder.AddChatCompletionService("chatcompletion", builder.Configuration["E2E_TEST_CHAT_COMPLETION_CACHE_DIR"]);
+builder.AddChatCompletionService("chatcompletion");
 builder.AddRedisClient("redis");
 
 JsonWebTokenHandler.DefaultInboundClaimTypeMap.Remove("sub");
